+++ histoire.setup.ts
@@ -2,6 +2,7 @@ import "./src/styles/miru.css";
 import "./.histoire/styles.css";
 import { defineSetupVue3 } from "@histoire/plugin-vue";
 
+import MiruPlugin from './src/miru';
 import FilesizeDirective from './src/directives/filesize';
 import OutsideDirective from "./src/directives/outside";
 import TooltipDirective from './src/directives/tooltip';
@@ -13,6 +14,9 @@ declare module "histoire" {
 }
 
 export const setupVue3 = defineSetupVue3(({ app, addWrapper }) => {
+    app.use(MiruPlugin, {
+        root: '#app'
+    });
     app.directive('click-outside', OutsideDirective);
     app.directive('filesize', FilesizeDirective);
     app.directive('tooltip', TooltipDirective);