+++ src/components/popover/PopoverStd.vue
@@ -10,17 +10,17 @@
 
         <slot name="popover" v-bind="props" :toggle="toggle" :show="show" :hide="hide" :visible="visible">
             <div class="popover-content">
-                <header v-if="$slots.header || props.title">
+                <header class="popover-header" v-if="$slots.header || props.title">
                     <slot name="header" v-bind="props" :toggle="toggle" :show="show" :hide="hide" :visible="visible">
                         <div class="popover-title">{{ props.title }}</div>
                     </slot>
                 </header>
     
-                <article v-if="$slots.content">
+                <article class="popover-article" v-if="$slots.content">
                     <slot name="content" v-bind="props" :toggle="toggle" :show="show" :hide="hide" :visible="visible"></slot>
                 </article>
                 
-                <footer v-if="$slots.footer">
+                <footer class="popover-footer" v-if="$slots.footer">
                     <slot name="footer" v-bind="props" :toggle="toggle" :show="show" :hide="hide" :visible="visible"></slot>
                 </footer>
             </div>
@@ -128,7 +128,7 @@ import { computed, ref, watch } from 'vue';
 
 // Define Component
 const props = withDefaults(defineProps<PopoverStdProps>(), {
-    placement: 'top',
+    placement: 'bottom',
     offset: 10,
     arrow: true
 });
@@ -217,18 +217,22 @@ defineExpose({
 <style scoped>
 .popover {
     --popover-background: theme(backgroundColor.white);
+    --popover-shadow: 212 212 216;
 }
 
 .popover {
-    @apply w-auto h-auto absolute flex flex-col opacity-0;
+    @apply w-auto h-auto max-w-64 absolute flex flex-col opacity-0 pointer-events-none;
     @apply duration-300 ease-in-out;
     z-index: 99;
     transition-property: opacity, margin;
     
     &.is-visible {
-        @apply opacity-100;
+        @apply opacity-100 pointer-events-auto;
     }
-    
+}
+
+/* Popover Position */
+.popover {
     &.popover-top,
     &.popover-top-start,
     &.popover-top-end {
@@ -270,61 +274,111 @@ defineExpose({
     }
 }
 
-.popover :slotted(.popover-content) {
-    @apply w-max flex text-xs font-semibold rounded-md;
-    @apply bg-white text-gray-800 shadow-gray-300 dark:bg-gray-100 dark:shadow-gray-900;
-    box-shadow: 0 0 0 1px var(--tw-shadow-color), 
-                0 4px 2px -2px var(--tw-shadow-color);
-}
-
-/** Arrow */
+/* Popover Arrow */
 .popover {
     & .popover-arrow {
-        @apply absolute w-0 h-0 border-solid;
-    }
-    
-    &.popover-top .popover-arrow,
-    &.popover-top-start .popover-arrow,
-    &.popover-top-end .popover-arrow {
-        @apply top-full;
-        border-width: 8px 8px 0 8px;
-        border-color: var(--popover-background) transparent transparent transparent;
+        @apply absolute border-solid;
+
+        &::before,
+        &::after {
+            @apply absolute top-0 left-0 inline-block w-0 h-0 border-solid border-transparent;
+            content: "";
+        }
     }
 
-    &.popover-right .popover-arrow,
-    &.popover-right-start .popover-arrow,
-    &.popover-right-end .popover-arrow {
-        @apply right-full;
-        border-width: 8px 8px 8px 0;
-        border-color: transparent var(--popover-background) transparent transparent;
-    }
-    &.popover-right-start .popover-arrow {
-        @apply -mt-1;
-    }
-    &.popover-right-end .popover-arrow {
-        @apply mt-1;
-    }
-
-    &.popover-bottom .popover-arrow,
-    &.popover-bottom-start .popover-arrow,
-    &.popover-bottom-end .popover-arrow {
-        @apply bottom-full;
-        border-width: 0 8px 8px 8px;
-        border-color: transparent transparent var(--popover-background) transparent;
+    &.popover-top,
+    &.popover-top-start,
+    &.popover-top-end {
+        & .popover-arrow {
+            @apply w-4 h-2 top-full;
+        }
+        & .popover-arrow::before {
+            @apply -mt-px -ml-px;
+            border-width: 10px 9px 0 9px;
+            border-top-color: rgb(var(--popover-shadow));
+        }
+        & .popover-arrow::after {
+            @apply -mt-px;
+            border-width: 8px 8px 0 8px;
+            border-top-color: var(--popover-background);
+        }
     }
     
-    &.popover-left .popover-arrow,
-    &.popover-left-start .popover-arrow,
-    &.popover-left-end .popover-arrow {
-        @apply left-full;
-        border-width: 8px 0 8px 8px;
-        border-color: transparent transparent transparent var(--popover-background);
+    &.popover-left,
+    &.popover-left-start,
+    &.popover-left-end {
+        & .popover-arrow {
+            @apply w-2 h-4 left-full;
+        }
+        & .popover-arrow::before {
+            @apply -mt-px -ml-px;
+            border-width: 9px 0 9px 10px;
+            border-left-color: rgb(var(--popover-shadow));
+        }
+        & .popover-arrow::after {
+            @apply -ml-px;
+            border-width: 8px 0 8px 8px;
+            border-left-color: var(--popover-background);
+        }
     }
-    &.popover-left-start .popover-arrow {
-        @apply -mt-1;
+    
+    &.popover-bottom,
+    &.popover-bottom-start,
+    &.popover-bottom-end {
+        & .popover-arrow {
+            @apply w-4 h-2 -translate-y-full;
+        }
+        & .popover-arrow::before {
+            @apply -mt-px -ml-px;
+            border-width: 0 9px 10px 9px;
+            border-bottom-color: rgb(var(--popover-shadow));
+        }
+        & .popover-arrow::after {
+            @apply mt-px;
+            border-width: 0 8px 8px 8px;
+            border-bottom-color: var(--popover-background);
+        }
     }
-    &.popover-left-end .popover-arrow {
-        @apply mt-1;
+    
+    &.popover-right,
+    &.popover-right-start,
+    &.popover-right-end {
+        & .popover-arrow {
+            @apply w-2 h-4 -translate-x-full;
+        }
+        & .popover-arrow::before {
+            @apply -ml-px -mt-px;
+            border-width: 9px 10px 9px 0;
+            border-right-color: rgb(var(--popover-shadow));
+        }
+        & .popover-arrow::after {
+            @apply ml-px;
+            border-width: 8px 8px 8px 0;
+            border-right-color: var(--popover-background);
+        }
     }
+}
+
+/* Popover Slots */
+.popover :slotted(.popover-content) {
+    @apply w-full flex flex-col rounded;
+    @apply bg-white text-gray-800 shadow-gray-300;
+    @apply dark:bg-gray-100 dark:shadow-gray-900;
+    box-shadow: 0 0 0 1px rgb(var(--popover-shadow)), 
+                0 4px 2px -2px rgb(var(--popover-shadow) / 0.5);
+}
+
+.popover :slotted(.popover-header) {
+    @apply px-3 pt-4 pb-1 border-b;
+    @apply border-gray-300;
+}
+
+.popover :slotted(.popover-title) {
+    @apply uppercase text-xs tracking-wider;
+    @apply text-gray-400 dark:text-gray-700;    
+}
+
+.popover :slotted(.popover-article) {
+    @apply p-3;
 }
 </style>