+++ src/miru.ts
@@ -1,8 +1,22 @@
-import type { Plugin } from "vue";
+import type { App, Plugin } from "vue";
 
 import FilesizeDirective from './directives/filesize';
 import OutsideDirective from "./directives/outside";
 import TooltipDirective from './directives/tooltip';
+
+export interface MiruConfig {
+    /**
+     * The application root selector (defaults to '#app').
+     */
+    root?: string;
+}
+
+/**
+ * Configuration Object
+ */
+const MiruOptions: Required<MiruConfig> = {
+    root: '#app'
+};
 
 /**
  * Miru Vue3 Plugin
@@ -13,19 +27,24 @@ const MiruPlugin = {
      * @param app 
      * @param options 
      */
-    install(app, ...options) {
+    install(app: App, options: MiruConfig) {
+        if (typeof options == 'object' && options !== null) {
+            for (const [key, val] of Object.entries(options)) {
+                if (typeof MiruOptions[(key as keyof MiruConfig)] != 'undefined') {
+                    MiruOptions[(key as keyof MiruConfig)] = val;
+                }
+            }
+        }
+
+        app.config.globalProperties.$miru = (key: keyof MiruConfig) => {
+            return MiruOptions[key] ?? null;
+        };
+
         app.directive('click-outside', OutsideDirective);
         app.directive('filesize', FilesizeDirective);
         app.directive('tooltip', TooltipDirective);
     },
-} as Plugin;
-
-/**
- * Setup Handler
- */
-function createMiru() {
-    return MiruPlugin
-}
+} as Plugin<MiruConfig>;
 
 // Export Module
-export { MiruPlugin, createMiru };
+export default MiruPlugin;